+++ tests/conftest.py
@@ -1,4 +1,5 @@
 import os
+import sys
 from pathlib import Path
 
 import pytest
@@ -10,6 +11,19 @@ def clean_environment():
             del os.environ[envvar]
         except KeyError:
             pass
+
+
+def update_environment():
+
+    # Set default Grafana version.
+    GRAFANA_VERSION_DEFAULT = "8.1.5"
+    if "GRAFANA_VERSION" not in os.environ:
+        os.environ["GRAFANA_VERSION"] = GRAFANA_VERSION_DEFAULT
+
+    # Report about Grafana version.
+    GRAFANA_VERSION = os.environ["GRAFANA_VERSION"]
+    sys.stderr.write(f"INFO: Running tests against Grafana version {GRAFANA_VERSION}\n")
+    sys.stderr.flush()
 
 
 @pytest.fixture(scope="session")
@@ -37,3 +51,4 @@ def docker_grafana(docker_services):
 
 
 clean_environment()
+update_environment()